author book integration

+++ books/src/main/java/pl/edu/pwr/ztw/books/service/AuthorsService.java
@@ -1,5 +1,6 @@
 package pl.edu.pwr.ztw.books.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.edu.pwr.ztw.books.model.Author;
 
@@ -8,6 +9,10 @@ import java.util.*;
 @Service
 public class AuthorsService implements IAuthorsService {
     private static List<Author> authorsRepo = new ArrayList<>();
+
+    @Autowired
+    private BooksService booksService;
+
     static {
         authorsRepo.add(new Author(1, "Henryk Sienkiewicz"));
         authorsRepo.add(new Author(2, "Stanisław Reymont"));
@@ -47,6 +52,9 @@ public class AuthorsService implements IAuthorsService {
 
     @Override
     public void deleteAuthor(int id) {
+        if (booksService.getBooks().stream().anyMatch(b -> b.getAuthorId() == id)) {
+            throw new IllegalArgumentException("Cannot delete author with existing books");
+        }
         if (!authorsRepo.removeIf(a -> a.getId() == id)) {
             throw new NoSuchElementException("Author not found with ID: " + id);
         }
